refactor(events): move events to seperate classes under `events/` folder
