+++ src/commands/slash/Main/hub/servers.ts
@@ -1,7 +1,7 @@
 import { colors, emojis } from '#main/utils/Constants.js';
 import db from '#main/utils/Db.js';
 import { t } from '#main/utils/Locale.js';
-import { Pagination } from '#main/utils/Pagination.js';
+import { Pagination } from '#main/modules/Pagination.js';
 import { resolveEval } from '#main/utils/Utils.js';
 import { type ChatInputCommandInteraction, EmbedBuilder } from 'discord.js';
 import Hub from './index.js';
